+++ bin/metrics-rabbitmq-exchange.rb
@@ -24,44 +24,10 @@
 # Released under the same terms as Sensu (the MIT license); see LICENSE
 # for details.
 
-require 'sensu-plugin/metric/cli'
-require 'socket'
-require 'carrot-top'
-require 'inifile'
+require 'sensu-plugins-rabbitmq'
 
 # main plugin class
-class RabbitMQExchangeMetrics < Sensu::Plugin::Metric::CLI::Graphite
-  option :host,
-         description: 'RabbitMQ management API host',
-         long: '--host HOST',
-         default: 'localhost'
-
-  option :port,
-         description: 'RabbitMQ management API port',
-         long: '--port PORT',
-         proc: proc(&:to_i),
-         default: 15_672
-
-  option :vhost,
-         description: 'Regular expression for filtering the RabbitMQ vhost',
-         short: '-v',
-         long: '--vhost VHOST'
-
-  option :username,
-         description: 'RabbitMQ management API user',
-         long: '--username USER',
-         default: 'guest'
-
-  option :password,
-         description: 'RabbitMQ management API password',
-         long: '--password PASSWORD',
-         default: 'guest'
-
-  option :scheme,
-         description: 'Metric naming scheme, text to prepend to $exchange_name.$metric',
-         long: '--scheme SCHEME',
-         default: "#{Socket.gethostname}.rabbitmq"
-
+class RabbitMQExchangeMetrics < Sensu::Plugin::RabbitMQ::Metrics
   option :filter,
          description: 'Regular expression for filtering exchanges',
          long: '--filter REGEX'
@@ -70,61 +36,9 @@ class RabbitMQExchangeMetrics < Sensu::Plugin::Metric::CLI::Graphite
          description: 'Regular expression for filtering metrics in each exchange',
          long: '--metrics REGEX'
 
-  option :ssl,
-         description: 'Enable SSL for connection to the API',
-         long: '--ssl',
-         boolean: true,
-         default: false
-
-  option :ini,
-         description: 'Configuration ini file',
-         short: '-i',
-         long: '--ini VALUE'
-
-  def acquire_rabbitmq_exchanges
-    begin
-      if config[:ini]
-        ini = IniFile.load(config[:ini])
-        section = ini['auth']
-        username = section['username']
-        password = section['password']
-      else
-        username = config[:username]
-        password = config[:password]
-      end
-
-      rabbitmq_info = CarrotTop.new(
-        host: config[:host],
-        port: config[:port],
-        user: username,
-        password: password,
-        ssl: config[:ssl]
-      )
-    rescue
-      warning 'could not get rabbitmq exchange info'
-    end
-
-    if config[:vhost]
-      return rabbitmq_info.exchanges.select { |x| x['vhost'].match(config[:vhost]) }
-    end
-
-    rabbitmq_info.exchanges
-  end
-
-  def dotted_keys(hash, prefix = '', keys = [])
-    hash.each do |k, v|
-      if v.is_a? Hash
-        keys = dotted_keys(v, prefix + k + '.', keys)
-      else
-        keys << prefix + k
-      end
-    end
-    keys
-  end
-
   def run
     timestamp = Time.now.to_i
-    acquire_rabbitmq_exchanges.each do |exchange|
+    acquire_rabbitmq_info(:exchanges).each do |exchange|
       if config[:filter]
         next unless exchange['name'].match(config[:filter])
       end